checkout

+++ database/migrations/2019_08_31_144036_create_orders_table.php
@@ -21,7 +21,7 @@ class CreateOrdersTable extends Migration
             $table->string('address');
             $table->string('email');
             $table->string('phone');
-            $table->decimal('money');
+            $table->float('money');
             $table->text('message')->nullable();
             $table->integer('status')->default(0);
             $table->timestamps();
